requirements.txt
accelerate>=0.34.0  # Also recommended to update accelerate for new PyTorch support
einops==0.8.0
h5py==3.11.0
matplotlib==3.9.0
mplhep==0.3.49
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
rootutils==1.0.7
SciencePlots==2.1.1
scipy==1.13.1
tensorboard==2.17.0
torch>=2.7.0
torchinfo>=1.8.0
torchsde==0.2.6
transformers==4.43.2
tqdm==4.66.4
wandb==0.17.0
timm>=1.0.0  # for mlp
hydra-core==1.3.2